Add WinBet.py
This is a Python script that uses Selenium to browse WinBet website and put bets on matches that are begin played at the moment. There is a pseudo-algorithm, which tries to guess which matches are good and which are bad. The algorithm doesn't work perfectly, as the data from WinBet isn't complete.